Reuse setTheme
